Make saving before merging with sfm branch
